// Frontend/podium-frontend/src/app/core/models/common.models.ts
export const Roles = {
  Student: 'Student',
  Guardian: 'Guardian',
  BandStaff: 'BandStaff',
  Director: 'Director',
  Admin: 'Admin'
} as const;

export type UserRole = typeof Roles[keyof typeof Roles];

export const Permissions = {
  ViewStudents: 'ViewStudents',
  RateStudents: 'RateStudents',
  ContactStudents: 'ContactStudents',
  SendOffers: 'SendOffers',
  ManageOffers: 'ManageOffers',
  ManageEvents: 'ManageEvents',
  ViewEvents: 'ViewEvents',
  ManageStaff: 'ManageStaff',
  ViewStaff: 'ViewStaff',
  ManageBand: 'ManageBand',
  ViewBandDetails: 'ViewBandDetails'
} as const;